Greyboxed Level Menu & Added Medal Sprites
Created a separate scene for the Level Menu. BGM and a different skybox still needs to be added. Interactivity also needs to be added.
There is an empty object called medals which inhabits the Medal image locations. The medal sprites are located under textures. The medal sprites should appear/ change based on the requirements we set.